# Conflicts:
#	base/core/include/motis/core/schedule/station.h
#	base/loader/include/motis/loader/build_stations.h
#	base/loader/include/motis/loader/loader_options.h
#	base/loader/src/build_stations.cc
#	base/loader/src/graph_builder.cc
#	modules/routing/include/motis/routing/output/label_chain_parser.h
#	modules/rt/include/motis/rt/additional_service_builder.h
#	modules/rt/include/motis/rt/separate_trip.h
#	modules/rt/src/rt_handler.cc
#	modules/tripbased/src/preprocessing.cc

+++ base/core/include/motis/core/access/realtime_access.h
@@ -26,7 +26,7 @@ ev_key const& get_current_ev_key(schedule const&, ev_key const&);
 
 ev_key const& get_orig_ev_key(schedule const&, ev_key const&);
 
-int get_schedule_track(schedule const&, ev_key const&);
+uint16_t get_schedule_track(schedule const&, ev_key const&);
 
 int get_schedule_track(schedule const&, edge const* route_edge,
                        light_connection const*, event_type);
